requirements.txt
numpy
matplotlib
pillow
torch
torchvision
tensorboard
tqdm